+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules.hs
@@ -15,6 +15,7 @@ module Cardano.Ledger.Conway.Rules (
   module Cardano.Ledger.Conway.Rules.Tickf,
   module Cardano.Ledger.Conway.Rules.Ratify,
   module Cardano.Ledger.Conway.Rules.Gov,
+  module Cardano.Ledger.Conway.Rules.Utxo,
   module Cardano.Ledger.Conway.Rules.Utxos,
   module Cardano.Ledger.Conway.Rules.Utxow,
 )
@@ -36,7 +37,7 @@ import Cardano.Ledger.Conway.Rules.NewEpoch
 import Cardano.Ledger.Conway.Rules.Pool ()
 import Cardano.Ledger.Conway.Rules.Ratify
 import Cardano.Ledger.Conway.Rules.Tickf
-import Cardano.Ledger.Conway.Rules.Utxo ()
+import Cardano.Ledger.Conway.Rules.Utxo
 import Cardano.Ledger.Conway.Rules.Utxos
 import Cardano.Ledger.Conway.Rules.Utxow
 import Cardano.Ledger.Shelley.Rules (ShelleyTickEvent (..))